+++ resources/schemas/index.js
@@ -1,7 +1,9 @@
 import createUserSchema from './create-user+v1';
 import loginUserSchema from './login-user+v1';
+import createChatSchema from './create-chat+v1';
 
 export default {
   createUserSchema,
   loginUserSchema,
+  createChatSchema,
 };