+++ src/func/chat/index.js
@@ -1,0 +1,5 @@
+import createChatController from './create';
+
+export {
+  createChatController,
+};