+++ src/routes.js
@@ -1,4 +1,5 @@
 import { login, register } from './func/auth';
+import { createChatController } from './func/chat';
 import { jwtAuthz } from './lib/middlewares';
 
 const routes = [
@@ -35,6 +36,15 @@ const routes = [
       },
     }),
     middlewares: [(req, res, next) => jwtAuthz(req, res, next)(['user:profile'])],
+    meta: {
+      isProtected: true,
+    },
+  },
+  {
+    path: '/chat',
+    method: 'POST',
+    controller: createChatController,
+    middlewares: [(req, res, next) => jwtAuthz(req, res, next)(['create:chat'])],
     meta: {
       isProtected: true,
     },