+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Added
 
+- Added `create-chat+v1.json` schema.
+- Added `createChatController` tp create a new chat.
 - Added initial Postman collection to test integration.
 - Added `scopes` in JWT claims to handle API Authorization.
 - Added `ForbiddenError` to handle API Authorization Error.